Keep file structure consistent with upstream type definition
