+++ types/guacamole-common-js/lib/WebSocketTunnel.d.ts
@@ -1,0 +1,11 @@
+import { Tunnel } from './Tunnel';
+
+/**
+ * Guacamole Tunnel implemented over WebSocket via XMLHttpRequest.
+ */
+export class WebSocketTunnel extends Tunnel {
+    /**
+     * @param tunnelURL The URL of the WebSocket tunneling service.
+     */
+    constructor(tunnelURL: string);
+}
