+++ types/guacamole-common-js/test/guacamole-common-js-test.ts
